+++ modules/kubernetes/ingress-nginx/main.tf
@@ -40,7 +40,7 @@ resource "helm_release" "ingress_nginx" {
   chart       = "ingress-nginx"
   name        = "ingress-nginx"
   namespace   = kubernetes_namespace.this.metadata[0].name
-  version     = "3.40.0"
+  version     = "4.0.17"
   max_history = 3
   values = [templatefile("${path.module}/templates/values.yaml.tpl", {
     http_snippet           = var.http_snippet
@@ -57,7 +57,7 @@ resource "helm_release" "ingress_nginx" {
     linkerd_enabled           = var.linkerd_enabled
     datadog_enabled           = var.datadog_enabled
     allow_snippet_annotations = var.allow_snippet_annotations
-    default_ingress_class     = false
+    default_ingress_class     = true
   })]
 }
 
@@ -72,7 +72,7 @@ resource "helm_release" "ingress_nginx_public" {
   chart       = "ingress-nginx"
   name        = "ingress-nginx-public"
   namespace   = kubernetes_namespace.this.metadata[0].name
-  version     = "3.40.0"
+  version     = "4.0.17"
   max_history = 3
   values = [templatefile("${path.module}/templates/values.yaml.tpl", {
     http_snippet           = var.http_snippet
@@ -104,7 +104,7 @@ resource "helm_release" "ingress_nginx_private" {
   chart       = "ingress-nginx"
   name        = "ingress-nginx-private"
   namespace   = kubernetes_namespace.this.metadata[0].name
-  version     = "3.40.0"
+  version     = "4.0.17"
   max_history = 3
   values = [templatefile("${path.module}/templates/values.yaml.tpl", {
     http_snippet           = var.http_snippet